Initial standart project layout
