+++ pkg/models/models.go
@@ -1,0 +1,9 @@
+package models
+
+type NumbersRequest struct {
+	Numbers []int `json:"numbers"`
+}
+
+type SumResponse struct {
+	Sum int `json:"sum"`
+}
